Apply dynamic imports to TimesSquareMainGitHubNav
- Created TimesSquareMainGitHubNavClient.tsx for client-only rendering
- Updated TimesSquareMainGitHubNav.tsx to use dynamic import with ssr:
  false
- Prevents useGitHubContentsListing SWR hook from executing during SSR
- Maintains full client-side navigation functionality

This resolves potential SSR issues with the GitHub contents listing hook
that fetches repository navigation data from the Times Square API.
